Added DIversityV2
This new diversity file is supposed to be cleaner and easier to run than
the original file that had more experimental work. I don't think I can
get it one click reproducible however due to the size of the files and
the difficulty of getting them out of the Government statistics
websites.